Lessons 34-37

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutation = {
   async createItem(parent, args, ctx, info) {
@@ -49,10 +50,16 @@ const Mutation = {
     const where = { id: args.id };
 
     //Find the Item
-    const item = await ctx.db.query.item({ where }, `{ id, title}`);
+    const item = await ctx.db.query.item({ where }, `{ id, title, user {id}}`);
 
     //Check if they own the item, or have the permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permissions to do that!");
+    }
 
     //Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
@@ -167,6 +174,40 @@ const Mutation = {
 
     // Return new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+
+    // query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+
+    // check the have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+
+    // update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
